v1.7.0 #1 Updated licenses to include Padre Build of Padre actions under OSX (needed to removed MB_ICONERROR) Added: SWS: Unselect all items/tracks/env points (depending on focus)   SWS: Unselect all items/tracks/env points

+++ Padre/padreEnvelopeProcessor.cpp
@@ -444,7 +444,7 @@ EnvelopeProcessor::ErrorCode EnvelopeProcessor::generateTakeLfo(MediaItem_Take* 
 	char* cNewState = new char[newState.size()];
 	strcpy(cNewState, newState.c_str());
 
-	if(!GetSetEnvelopeState(envelope, cNewState, newState.size()))
+	if(!GetSetEnvelopeState(envelope, cNewState, (int)newState.size()))
 		return eERRORCODE_UNKNOWN;
 
 //! \bug I can't/don't have to delete cNewState ???
